requirements.txt
python-telegram-bot[job-queue]==21.10
httpx
